requirements.txt
bertopic
keybert
numpy
pandas
faiss-cpu
fastapi
umap-learn
nltk
seaborn
matplot
sentence-transformers
psycopg2-binary
google-cloud-storage
pyarrow
uvicorn
